docs: move to typedoc 0.20
