+++ package.json
@@ -17,18 +17,16 @@
     "prettier": "^2.0.5",
     "sinon": "^9.0.1",
     "ts-node": "^8.1.0",
+    "typedoc": "^0.21.4",
     "typedoc-loopingz-theme": "^1.0.0",
-    "typedoc-plugin-lerna-packages": "^0.3.1",
-    "typedoc-plugin-mermaid": "^1.3.0",
-    "typedoc-plugin-nojekyll": "^1.0.1",
     "typedoc-plugin-loopingz-pages": ">=0.1.0",
+    "typedoc-plugin-mermaid": "^1.4.0",
     "typescript": "^4.0.0"
   },
   "author": "loopingz",
   "license": "MIT",
   "resolutions": {
-    "@types/node": "^12.0.0",
-    "@types/mime": "^2.0.3"
+    "@types/node": "^12.0.0"
   },
   "bugs": {
     "url": "https://github.com/loopingz/webda.io/issues"
@@ -49,7 +47,7 @@
     "lerna:publish": "lerna publish",
     "docs": "yarn docs:typedoc",
     "docs:publish": "yarn docs:typedoc && gh-pages -t -d docs/typedoc",
-    "docs:typedoc": "typedoc --out docs/typedoc ."
+    "docs:typedoc": "typedoc --packages . --out docs/typedoc"
   },
   "command": {
     "publish": {
