+++ packages/core/package.json
@@ -33,7 +33,6 @@
     "jsonpath": "^1.0.1",
     "jsonwebtoken": "^8.5.1",
     "merge": "^2.1.1",
-    "mime": "^2.4.2",
     "mime-types": "^2.1.24",
     "mustache": "^3.0.1",
     "node-cron": "^2.0.3",
